* 'verify_inflight' of https://github.com/noclip-code/fio:
  verify: add versioning to verify_header
  verify: clear inflight log in between loops
  verify: plumb inflight write information through verify state
  verify: rework write tracking for use with verify_save_state()
  verify: make numberio uint64_t

+++ iolog.h
@@ -260,7 +260,7 @@ struct io_piece {
 		struct fio_file *file;
 	};
 	unsigned long long offset;
-	unsigned short numberio;
+	uint64_t numberio;
 	unsigned long len;
 	unsigned int flags;
 	enum fio_ddir ddir;
